requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
yfinance>=0.2.40
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0
diskcache>=5.6.3
httpx>=0.27.0
pytest>=8.0.0
